Adjust parallelization to avoid freezing
The maximum amount of parallel processes is set to be half the amount of available CPUs. Furthermore a timeout condition for the subprocesses is added. Right now a subprocess is terminated after 60 minutes. This should help when processing large datasets and you want to make sure it does not get stuck.